refactor: test Input 컴포넌트를 기존 Input 컴포넌트로 변경
