+++ src/components/basic/Input/index.jsx
@@ -1,61 +1,46 @@
 import { useCallback } from 'react';
 import styled from '@emotion/styled';
-import PropTypes, { number } from 'prop-types';
+import PropTypes from 'prop-types';
 import colors from 'utils/constants/colors';
-import Icon from 'components/basic/Icon';
 
-const { mainGreen, mainRed, grayLight } = colors;
+
+const { black, mainRed, grayLight } = colors;
 
 const InputWrapper = styled.div`
-  width: ${({ width }) => (typeof width === number ? width : `${width}px`)};
-  padding: 22px 22px;
-  box-sizing: border-box;
-  position: relative;
+  display: ${({ block }) => (block ? 'block' : 'inline-block')};
+  width: ${({ width }) => (width ? `${width}px` : '100%')};
 `;
 
-const Label = styled.label`
-  display: ${({ isLabel }) => (isLabel ? 'block' : 'none')};
+const Label = styled.div`
   font-size: 20px;
-  margin: 0 0 15px 0;
+  margin-bottom: 5px;
   font-weight: 500;
   text-align: left;
-  position: relative;
-  left: 5px;
-`;
-
-const SearchIcon = styled(Icon)`
-  display: ${({ isIcon }) => (isIcon ? 'block' : 'none')};
-  position: absolute;
-  left: 40px;
-  bottom: 50%;
-  top: 50%;
-  margin: auto 0 auto 0;
 `;
 
 const StyledInput = styled.input`
   ${({ inputStyle }) => inputStyle};
-  padding: 0 ${({ isIcon }) => (isIcon ? '50px' : '22px')};
   width: 100%;
-  position: relative;
-  border: none;
+  padding: 22px 22px;
+  border-radius: 15px;
   box-sizing: border-box;
-  border-radius: 10px;
-  border: ${({ inValid }) =>
-    inValid ? `1px solid ${mainRed}` : `1px solid ${grayLight}`};
+  border: 1px solid ${({inValid }) => inValid ? mainRed : grayLight };
+  
+  ::placeholder {
+    color: #a3a3a3;
+  }
 
   &:focus {
-    outline-color: ${mainGreen};
+    outline-color: ${black};
   }
 `;
 
-const Input = ({
+const NewInput = ({
+  block,
+  width,
   type = 'text',
   name = 'name',
-  width = '100%',
-  height = '70px',
-  label,
-  isIcon = false,
-  isLabel = false,
+  label = 'label',
   inValid = false,
   required = false,
   fontSize = '18px',
@@ -63,9 +48,9 @@ const Input = ({
   onChange,
   ...props
 }) => {
+
   const InputStyle = {
     fontSize,
-    height,
   };
 
   const handleChange = useCallback(
@@ -76,35 +61,29 @@ const Input = ({
   );
 
   return (
-    <InputWrapper width={width}>
-      <Label isLabel={isLabel}>{label}</Label>
-      <StyledInput
-        type={type}
-        name={name}
-        height={height}
-        isIcon={isIcon}
-        required={required}
-        placeholder={placeholder}
-        inputStyle={InputStyle}
-        onChange={handleChange}
-        inValid={inValid}
-      />
-      <SearchIcon isIcon={isIcon} name="LIKE_ICON"></SearchIcon>
-    </InputWrapper>
+    <InputWrapper block={block} width={width}>
+      {label && <Label>{label}</Label>}
+        <StyledInput
+          type={type}
+          name={name}
+          required={required}
+          placeholder={placeholder}
+          inputStyle={InputStyle}
+          onChange={handleChange}
+          inValid={inValid}
+        />
+      </InputWrapper>
   );
 };
 
-export default Input;
+export default NewInput;
 
-Input.propTypes = {
+NewInput.propTypes = {
+  block: PropTypes.bool,
+  width: PropTypes.string,
   type: PropTypes.string,
   name: PropTypes.string,
-  width: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  height: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
-  initialValue: PropTypes.string,
   label: PropTypes.string,
-  isIcon: PropTypes.bool,
-  isLabel: PropTypes.bool,
   inValid: PropTypes.bool,
   required: PropTypes.bool,
   fontSize: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
